Gen_AI_1/requirements.txt
openai
groq
google-generativeai
moonshotai

langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
langchain-google-genai
langchain-moonshot

faiss-cpu
chromadb
pinecone-client

tiktoken
python-dotenv

requests
beautifulsoup4
lxml

fastapi
uvicorn
pydantic

streamlit

pandas
numpy

langchain-groq 
langchain-core 
langchain-community